+++ src/main/java/org/chzz/market/domain/notification/service/NotificationService.java
@@ -3,30 +3,20 @@ package org.chzz.market.domain.notification.service;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.IOException;
 import java.time.Instant;
-import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.chzz.market.domain.notification.dto.NotificationRealMessage;
 import org.chzz.market.domain.notification.dto.response.NotificationResponse;
 import org.chzz.market.domain.notification.dto.response.NotificationSseResponse;
 import org.chzz.market.domain.notification.entity.Notification;
 import org.chzz.market.domain.notification.error.NotificationErrorCode;
 import org.chzz.market.domain.notification.error.NotificationException;
-import org.chzz.market.domain.notification.event.NotificationEvent;
 import org.chzz.market.domain.notification.repository.EmitterRepositoryImpl;
 import org.chzz.market.domain.notification.repository.NotificationRepository;
-import org.chzz.market.domain.user.entity.User;
-import org.chzz.market.domain.user.repository.UserRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.transaction.event.TransactionalEventListener;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 @Service
@@ -34,10 +24,8 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 @Transactional(readOnly = true)
 @Slf4j
 public class NotificationService {
-    private final RedisPublisher redisPublisher;
     private final NotificationRepository notificationRepository;
     private final EmitterRepositoryImpl emitterRepository;
-    private final UserRepository userRepository;
     private final ObjectMapper objectMapper;
 
     /**
@@ -52,31 +40,6 @@ public class NotificationService {
         setupEmitterCallbacks(userId, emitter);
         sendInitialConnectionEvent(userId, emitter);
         return emitter;
-    }
-
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    @TransactionalEventListener
-    public void sendNotification(final NotificationEvent notificationEvent) {
-        // 이벤트 수신 로그
-        log.info("알림 이벤트 수신 - Type: {}, Message: '{}', User IDs: {}, Image: {}",
-                notificationEvent.type(),
-                notificationEvent.message(),
-                notificationEvent.userIds(),
-                notificationEvent.image() != null ? notificationEvent.image().getId() : "No Image");
-        // 1. 알림 메시지 저장
-        Map<Long, User> userMap = userRepository.findAllById(notificationEvent.userIds())
-                .stream()
-                .collect(Collectors.toMap(User::getId, Function.identity()));
-        List<Notification> notifications = createNotifications(notificationEvent, userMap);
-        notificationRepository.saveAll(notifications);
-
-        // 2. 사용자 ID와 알림 ID를 매핑하는 맵 생성과 메시지 발행
-        Map<Long, Long> userNotificationMap = notifications.stream()
-                .collect(Collectors.toMap(n -> n.getUser().getId(), Notification::getId));
-
-        // 3. Redis에 메시지 발행
-        redisPublisher.publish(new NotificationRealMessage(userNotificationMap, notificationEvent.message(),
-                notificationEvent.type()));
     }
 
     /**
@@ -152,41 +115,6 @@ public class NotificationService {
         } catch (Exception e) {
             log.error("Error sending initial connection event to user {}", userId);
         }
-    }
-
-    /**
-     * 알림 객체 목록을 생성합니다.
-     *
-     * @param notificationEvent 알림 이벤트 데이터
-     * @param userMap           사용자 ID와 사용자 객체의 매핑
-     * @return 생성된 알림 객체 목록
-     */
-    private List<Notification> createNotifications(NotificationEvent notificationEvent,
-                                                   Map<Long, User> userMap) {
-        return notificationEvent.userIds().stream()
-                .map(userId -> createNotification(notificationEvent, userMap.get(userId)))
-                .flatMap(Optional::stream)
-                .toList();
-    }
-
-    /**
-     * 알림 객체를 생성합니다.
-     *
-     * @param notificationEvent 알림 이벤트 데이터
-     * @param user              사용자 객체
-     * @return Optional로 감싼 알림 객체
-     */
-    private Optional<Notification> createNotification(NotificationEvent notificationEvent, User user) {
-        if (user == null) {
-            return Optional.empty();
-        }
-
-        return Optional.of(Notification.builder()
-                .message(notificationEvent.message())
-                .user(user)
-                .image(notificationEvent.image())
-                .type(notificationEvent.type())
-                .build());
     }
 
     private Notification findNotificationByUserAndId(Long userId, Long notificationId) {
